feat(optimize): tag components rewrite
